requirements.txt
fastapi==0.116.1
pydantic==2.11.7
uvicorn==0.35.0
sentence-transformers==5.1.0
torch==2.8.0
transformers==4.55.2
language-tool-python==2.9.4
nltk==3.9.1
pyspellchecker==0.7.8
